refactor : play, ladder
