+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -1,10 +1,5 @@
 package nextstep.ladder.domain;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.stream.Collectors;
-
 /**
  * @author han
  */
@@ -28,44 +23,5 @@ public class Ladder {
 
     public Lines getLines() {
         return lines;
-    }
-
-    public Play toPlay(Results results) {
-        Line line = this.lines.getLines().get(0);
-        List<Position> positions = first(line);
-
-        for (int i = 1; i < this.lines.getLines().size(); i++) {
-            Line line1 = this.lines.getLines().get(i);
-            positions = next(positions, line1);
-        }
-
-        return new Play(positions, results);
-    }
-
-    private List<Position> first(Line line) {
-        List<Position> positions = new ArrayList<>();
-
-        for (int i = 0; i < line.getPoints().size(); i++) {
-            Point point = line.getPoints().get(i);
-            positions.add(new Position(i, users.getUsers().get(i)).move(point));
-        }
-
-        return positions.stream()
-            .sorted(Comparator.comparing(Position::getIndex))
-            .collect(Collectors.toList());
-    }
-
-    private List<Position> next(List<Position> positions, Line line) {
-        List<Position> result = new ArrayList<>();
-
-        for (int i = 0; i < line.getPoints().size(); i++) {
-            Point point = line.getPoints().get(i);
-            Position position = positions.get(i);
-            result.add(Position.of(position).move(point));
-        }
-
-        return result.stream()
-            .sorted(Comparator.comparing(Position::getIndex))
-            .collect(Collectors.toList());
     }
 }
